straight forward conftest
