+++ .pytest.ini
@@ -3,4 +3,5 @@ addopts = --cov=ahriman --cov-report=term-missing:skip-covered --no-cov-on-fail 
 asyncio_default_fixture_loop_scope = function
 asyncio_mode = auto
 pythonpath = tests
+resource-path.directory-name-test-resources = ../../tests/testresources
 spec_test_format = {result} {docstring_summary}
